fix: Fix code models The codes in services.users and services.cats were realocated to models. users and models.cats, the correct files.
